feat(ui): use primevue DataTable
