+++ resources/ui/pages/About.vue
@@ -1,12 +1,8 @@
 <script setup>
-import ListItems from '../components/ListItems.vue'
-import Button from 'primevue/button'
 import Table from '../components/Table.vue'
 </script>
 
 <template>
   <h2>about page</h2>
-  <Button label="Submit" />
-  <ListItems />
   <Table />
 </template>
